Revert "Skip processing if buffer not modified"
This reverts commit f570656a6219591526c52a91b1ba1942a8184d59.

Turns out this broke things: if paren mode isn't run before other changes
are attempted, the other changes are processed in smart mode, and the
parenthesis/indent is broken, messing up the file.

The future solution is to delay paren mode until a change is made, but
it still must be run before the first change is applied to avoid breaking
the file.

See #98